+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/blocks/HologramProjector.java
@@ -1,17 +1,24 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.blocks;
 
+import java.util.List;
+
+import javax.annotation.Nonnull;
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
+import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
 import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
 import io.github.thebusybiscuit.cscorelib2.math.DoubleHandler;
+import io.github.thebusybiscuit.slimefun4.api.events.AndroidMineEvent;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockBreakHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockUseHandler;
@@ -29,12 +36,14 @@ public class HologramProjector extends SlimefunItem {
 
     private static final String OFFSET_PARAMETER = "offset";
 
+    @ParametersAreNonnullByDefault
     public HologramProjector(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe, ItemStack recipeOutput) {
         super(category, item, recipeType, recipe, recipeOutput);
 
         addItemHandler(onPlace(), onRightClick(), onBreak());
     }
 
+    @Nonnull
     private BlockPlaceHandler onPlace() {
         return new BlockPlaceHandler(false) {
 
@@ -51,13 +60,28 @@ public class HologramProjector extends SlimefunItem {
         };
     }
 
+    @Nonnull
     private BlockBreakHandler onBreak() {
-        return (e, item, fortune, drops) -> {
-            remove(e.getBlock());
-            return true;
+        return new BlockBreakHandler(true, true) {
+
+            @Override
+            public void onPlayerBreak(BlockBreakEvent e, ItemStack item, List<ItemStack> drops) {
+                remove(e.getBlock());
+            }
+
+            @Override
+            public void onExplode(Block b, List<ItemStack> drops) {
+                remove(b);
+            }
+
+            @Override
+            public void onAndroidBreak(AndroidMineEvent e) {
+                remove(e.getBlock());
+            }
         };
     }
 
+    @Nonnull
     public BlockUseHandler onRightClick() {
         return e -> {
             e.cancel();
@@ -71,7 +95,7 @@ public class HologramProjector extends SlimefunItem {
         };
     }
 
-    private static void openEditor(Player p, Block projector) {
+    private static void openEditor(@Nonnull Player p, @Nonnull Block projector) {
         ChestMenu menu = new ChestMenu(SlimefunPlugin.getLocalization().getMessage(p, "machines.HOLOGRAM_PROJECTOR.inventory-title"));
 
         menu.addItem(0, new CustomItem(Material.NAME_TAG, "&7Text &e(Click to edit)", "", "&r" + ChatColors.color(BlockStorage.getLocationInfo(projector.getLocation(), "text"))));
@@ -104,7 +128,7 @@ public class HologramProjector extends SlimefunItem {
         menu.open(p);
     }
 
-    private static ArmorStand getArmorStand(Block projector, boolean createIfNoneExists) {
+    private static ArmorStand getArmorStand(@Nonnull Block projector, boolean createIfNoneExists) {
         String nametag = BlockStorage.getLocationInfo(projector.getLocation(), "text");
         double offset = Double.parseDouble(BlockStorage.getLocationInfo(projector.getLocation(), OFFSET_PARAMETER));
         Location l = new Location(projector.getWorld(), projector.getX() + 0.5, projector.getY() + offset, projector.getZ() + 0.5);
@@ -128,7 +152,7 @@ public class HologramProjector extends SlimefunItem {
         return hologram;
     }
 
-    private static void remove(Block b) {
+    private static void remove(@Nonnull Block b) {
         ArmorStand hologram = getArmorStand(b, false);
 
         if (hologram != null) {